Update logo and refine doc labels and titles
This commit deletes the TrackDayR-logotype-light.svg logo and replaces it with TrackDayR-mark-light.svg for a cleaner branding.

In addition, it enhances documentation by updating titles in the bike and rider skinning sections from "Bike Skin WIKI" to "Bike Skinning" and "Rider Skin WIKI" to "Rider Skinning" respectively, for better brevity and understanding.

It also adds the 'Experimental' badge next to API references in the sidebar to notify users about the experimental status of the API.

All of these changes are made to improve branding visibility and make documentation more user-friendly and informative.